修复无法解析数字超过 u8 范围 Beats 的 Bug

+++ src/insn/directives_ty.rs
@@ -5,6 +5,6 @@ pub struct BpmParams {
 
 #[derive(Copy, Clone, PartialEq, Debug)]
 pub enum BeatDivisorParams {
-    NewDivisor(u8),
+    NewDivisor(u32),
     NewAbsoluteDuration(f32),
 }
